crud/requirements.txt
boto3==1.20.33
Flask==2.0.2
mysql_connector_repackaged==0.3.1
pandas==1.3.5
python-dotenv==0.19.2
